# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nula Logistics API documentation",
        "description": "API documentation for Products, Suppliers and Warehouses management. All routes start with /api namespace",
        "contact": {
            "email": "katerinka.melnichuk@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/warehouses/{warehouseId}/products": {
            "patch": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Assign a Product to a Warehouse",
                "operationId": "d15de4b047c574fb56d1421bdb3b500d",
                "parameters": [
                    {
                        "name": "warehouseId",
                        "in": "path",
                        "description": "ID of the Warehouse",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "ID of the Product",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Quantity of the Product",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product assigned to Warehouse successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}